sort by max

+++ AlmutalCore/Models/Node.cs
@@ -8,10 +8,10 @@ namespace AlmutalCore.Models
     {
         public Node RightNode { get; set; }
         public Node BottomNode { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Width { get; set; }
-        public float Length { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Width { get; set; }
+        public double Length { get; set; }
         public bool IsOccupied{ get; set; }
         public bool rotated{ get; set; }
         public int Id{ get; set; }
